pyproject.toml
[project]
name = "leadflow-ai"
version = "0.1.0"
description = "Full-stack AI platform for marketing and lead gen agencies"
requires-python = ">=3.13"

dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    # Cache
    "redis>=5.2.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Config
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    # HTTP client
    "httpx>=0.27.0",
    "bcrypt==4.0.1",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langchain-community>=0.4.1",
    "chromadb>=1.5.5",
    "pymupdf>=1.27.2.2",
    "tiktoken>=0.12.0",
    "langchain-text-splitters>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["leadflow_ai"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
